onie: Add links and link-count property
The links property is an array of references to all of the devices
associated with the Tomahawk that are links to the CPU.

The link-count property is an integer number of interfaces. This
could also be derived from the size of other proprties, but exists
for consistency checking and pre-sizing dynamic structures.

Signed-off-by: Kevin Paul Herbert <kph@platinasystems.com>